refs Anatoli Tsoi moved the user data to separated directory test-data
